SimCompRisk {prodlim}	R Documentation

Simulate competing risks data

Description

Simulate right censored competing risks data with two covariates X1 and X2. Both covariates have effect exp(1) on the hazards of event 1 and zero effect on the hazard of event 2.

Usage

SimCompRisk(N, ...)

Arguments

N
sample size

...
do nothing.

Details

This function calls crModel, then adds covariates and finally calls sim.lvm.

Value

data.frame with simulated data

Author(s)

Thomas Alexander Gerds

Examples

SimCompRisk(10)
            

[Package prodlim version 2018.04.18 Index]