SimSurv {prodlim}	R Documentation

Simulate survival data

Description

Simulate right censored survival data with two covariates X1 and X2, both have effect exp(1) on the hazard of the unobserved event time.

Usage

SimSurv(N, ...)

Arguments

N
sample size

...
do nothing

Details

This function calls survModel, then adds covariates and finally calls sim.lvm.

Value

data.frame with simulated data

Author(s)

Thomas Alexander Gerds

References

Bender, Augustin & Blettner. Generating survival times to simulate Cox proportional hazards models. Statistics in Medicine, 24: 1713-1723, 2005.

Examples

SimSurv(10)

[Package prodlim version 2018.04.18 Index]