as_tibble.prodlim {prodlim}	R Documentation

Turn prodlim objects into a tibble

Description

Collect results of a fitted prodlim object in a tibble

Usage

as_tibble.prodlim(x, ...)

Arguments

x
object obtained with function prodlim

...
passed to summary.prodlim

Details

By default object contains results for all fitted time points and all strata. Use arguments times and newdata of summary.prodlim to subset.

Value

A data.table with the results of the prodlim object

Author(s)

Thomas A. Gerds <tag@biostat.ku.dk>

See Also

prodlim, tibble

Examples

library(tibble)
set.seed(8)
d <- SimCompRisk(17)
fit <- prodlim(Hist(time,event)~X1,data=d)
tibble::as_tibble(fit)

[Package prodlim version 2026.03.11 Index]