checkCauses {prodlim}	R Documentation

Check availability of a cause in competing risk settings

Description

For competing risk settings, check if the requested cause is known to the object

Usage

checkCauses(cause, object)

Arguments

cause
cause of interest

object
object either obtained with Hist or prodlim

[Package prodlim version 2026.03.11 Index]