clusterTestData {prodlim}	R Documentation

Data set with a clustered data structure, for testing the product limit method.

Description

Data set with a clustered data structure, for testing the product limit method.

Format

A data frame with 42 observations on the following 4 variables.

list("midtimeX")

event time

list("eventX")

event status

list("patientid")

patient number

list("AnyCrownFracture")

a binary predictor

Examples

data(clusterTestData)
## maybe str(clusterTestData) ; plot(clusterTestData) ...

[Package prodlim version 1.5.7 Index]