crModel {prodlim}	R Documentation

Competing risks model for simulation

Description

Competing risks model for simulation

Usage

crModel()

Details

Create a competing risks model with to causes to simulate a right censored event time data without covariates

This function requires the lava package.

Value

A structural equation model initialized with four variables: the latent event times of two causes, the latent right censored time, and the observed right censored event time.

Author(s)

Thomas A. Gerds

Examples

library(lava)
m <- crModel()
d <- sim(m,6)
print(d)

[Package prodlim version 2018.04.18 Index]