dimColor {prodlim}	R Documentation

Dim a given color to a specified density

Description

This function calls first col2rgb on a color name and then uses rgb to adjust the intensity of the result.

Usage

dimColor(col, density = 55)

Arguments

col
Color name or number passed to col2rgb.

density
Integer value passed as alpha coefficient to rgb between 0 and 255

Value

A character vector with the color code. See rgb for details.

Author(s)

Thomas A. Gerds <tag@biostat.ku.dk>

See Also

rgb col2rgb

Examples

dimColor(2,33)
dimColor("green",133)

[Package prodlim version 2018.04.18 Index]