followup {prodlim}	R Documentation

Estimation of the median potential followup time.

Description

The reverse Kaplan-Meier method estimates the median potential followup time.

Usage

followup(formula, cens.code = 0, data, ...)

Arguments

formula
A formula whose left hand side is a Hist or a Surv object specifying the event time and the event type where by default 0=censored, 1=event, 2=competing risk (if any). Use cens.code to change the value for censored.

cens.code
Value of the event

data
A data.frame in which all the variables of formula can be interpreted.

...
Arguments passed to prodlim.quantile.

Details

This is merely a wrapper for prodlim with argument reverse.

Value

The estimated median potential followup time with inter quartile ranges.

Author(s)

Thomas A. Gerds <tag@biostat.ku.dk>

References

Michael Schemper and Terry L. Smith. A note on quantifying follow-up in studies of failure time. Controlled Clinical Trials, 17(4):343–346, 1996.

See Also

prodlim

Examples

set.seed(8)
d <- SimCompRisk(117)

# overall
followup(Hist(time,event)~1,data=d)

# in strata defined by variable X1 
followup(Hist(time,event)~X1,data=d)

[Package prodlim version 2026.03.11 Index]