getEvent {prodlim}	R Documentation

Extract a column from an event history object.

Description

Extract a column from an event history object, as obtained with the function Hist.

Usage

getEvent(object, mode = "factor", column = "event")

Arguments

object
Object of class "Hist".

mode
Return mode. One of "numeric", "character", or "factor".

column
Name of the column to extract from the object.

Details

Since objects of class "Hist" are also matrices, all columns are numeric or integer valued. To extract a correctly labeled version, the attribute states of the object is used to generate factor levels.

Author(s)

Thomas Alexander Gerds <tag@biostat.ku.dk>

See Also

Hist

Examples

  dat= data.frame(time=1:5,event=letters[1:5])
  x=with(dat,Hist(time,event))
  ## inside integer
  unclass(x)
  ## extract event (the extra level "unknown" is for censored data)
  getEvent(x)

[Package prodlim version 2018.04.18 Index]