getStates {prodlim}	R Documentation

States of a multi-state model

Description

Extract the states of a multi-state model

Usage

getStates(object, ...)

Arguments

object
Object of class prodlim or Hist .

...
not used

Details

Applying this function to the fit of prodlim means to apply it to fit$model.response.

Value

A character vector with the states of the model.

Author(s)

Thomas A. Gerds

[Package prodlim version 2018.04.18 Index]