jackknife {prodlim}	R Documentation

Compute jackknife pseudo values.

Description

Compute jackknife pseudo values.

Usage

jackknife(object, times, cause, keepResponse = FALSE, ...)

Arguments

object
Object of class "prodlim".

times
Time points at which to compute pseudo values.

cause
Character (other classes are converted with as.character). For competing risks the cause of failure.

keepResponse
If TRUE add the model response, i.e. event time, event status, etc. to the result.

...
not used

Details

Compute jackknife pseudo values based on marginal Kaplan-Meier estimate of survival, or based on marginal Aalen-Johansen estimate of cumulative incidence.

Note

The R-package pseudo does a similar job, and appears to be a little faster in small samples, but much slower in large samples. See examples.

Author(s)

Thomas Alexander Gerds <tag@biostat.ku.dk>

References

Andersen PK & Perme MP (2010). Pseudo-observations in survival analysis Statistical Methods in Medical Research, 19(1), 71-99.

See Also

prodlim

Examples

## pseudo-values for survival models

d=SimSurv(20) 
f=prodlim(Hist(time,status)~1,data=d) 
jackknife(f,times=c(3,5))

## in some situations it may be useful to attach the
## the event time history to the result
jackknife(f,times=c(3,5),keepResponse=TRUE)

# pseudo-values for competing risk models
d=SimCompRisk(10) 
f=prodlim(Hist(time,event)~1,data=d) 
jackknife(f,times=c(3,10),cause=1)
jackknife(f,times=c(3,10,17),cause=2)

[Package prodlim version 2018.04.18 Index]