meanNeighbors {prodlim}	R Documentation

Helper function to obtain running means for prodlim objects.

Description

Compute average values of a variable according to neighborhoods.

Usage

meanNeighbors(x, y, ...)

Arguments

x
Object of class "neighborhood".

y
Vector of numeric values.

...
Not used.

Author(s)

Thomas Alexander Gerds <tag@biostat.ku.dk>

See Also

neighborhood

Examples

meanNeighbors(x=1:10,y=c(1,10,100,1000,1001,1001,1001,1002,1002,1002))

[Package prodlim version 2018.04.18 Index]