neighborhood {prodlim}	R Documentation

Nearest neighborhoods for kernel smoothing

Description

Nearest neighborhoods for the values of a continuous predictor. The result is used for the conditional Kaplan-Meier estimator and other conditional product limit estimators.

Usage

neighborhood(x, bandwidth = NULL, kernel = "box")

Arguments

x
Numeric vector – typically the observations of a continuous random variate.

bandwidth
Controls the distance between neighbors in a neighborhood. It can be a decimal, i.e.\ the bandwidth, or the string ‘"smooth"’, in which case N^{-1/4} is used, N being the sample size, or NULL in which case the dpik function of the package KernSmooth is used to find the optimal bandwidth.

kernel
Only the rectangular kernel ("box") is implemented.

Value

An object of class 'neighborhood'. The value is a list that includes the unique values of ‘x’ (values) for which a neighborhood, consisting of the nearest neighbors, is defined by the first neighbor (first.nbh) of the usually very long vector neighbors and the size of the neighborhood (size.nbh).

Further values are the arguments bandwidth, kernel, the total sample size n and the number of unique values nu.

Author(s)

Thomas Gerds

References

Stute, W. "Asymptotic Normality of Nearest Neighbor Regression Function Estimates", The Annals of Statistics, 1984,12,917–926.

See Also

dpik, prodlim

Examples

d <- SimSurv(20)
neighborhood(d$X2)

[Package prodlim version 2018.04.18 Index]