plotIllnessDeathModel {prodlim}	R Documentation

Plotting an illness-death-model.

Description

Plotting an illness-death-model using plot.Hist.

Usage

plotIllnessDeathModel(stateLabels, style = 1, recovery = FALSE, ...)

Arguments

stateLabels
Labels for the three boxes.

style
Either 1 or anything else, switches the orientation of the graph. Hard to explain in words, see examples.

recovery
Logical. If TRUE there will be an arrow from the illness state to the initial state.

...
Arguments passed to plot.Hist.

Author(s)

Thomas Alexander Gerds <tag@biostat.ku.dk>

See Also

plotCompetingRiskModel, plot.Hist

Examples

plotIllnessDeathModel()
plotIllnessDeathModel(style=2)
plotIllnessDeathModel(style=2,
                      stateLabels=c("a","b\nc","d"),
                      box1.col="yellow",
                      box2.col="green",
                      box3.col="red")

[Package prodlim version 2018.04.18 Index]