predict.prodlim {prodlim}	R Documentation

Predicting event probabilities from product limit estimates

Description

Evaluation of estimated survival or event probabilities at given times and covariate constellations.

Usage

## S3 method for class 'prodlim'
predict(object, times, newdata, level.chaos = 1,
  type = c("surv", "cuminc", "list"), mode = "list", bytime = FALSE,
  cause, ...)

Arguments

object
A fitted object of class "prodlim".

times
Vector of times at which to return the estimated probabilities (survival or absolute event risks).

newdata
A data frame with the same variable names as those that appear on the right hand side of the 'prodlim' formula. If there are covariates this argument is required.

level.chaos
Integer specifying the sorting of the output: ‘0’ sort by time and newdata; ‘1’ only by time; ‘2’ no sorting at all

type
Choice between "surv","cuminc","list":

"surv": predict survival probabilities only survival models

"cuminc": predict cumulative incidences only competing risk models

"list": find the indices corresponding to times and newdata. See value.

Defaults to "surv" for two-state models and to "cuminc" for competing risk models.

mode
Only for type=="surv" and type=="cuminc". Can either be "list" or "matrix". For "matrix" the predicted probabilities will be returned in matrix form.

bytime
Logical. If TRUE and mode=="matrix" the matrix with predicted probabilities will have a column for each time and a row for each newdata. Only when object$covariate.type>1 and more than one time is given.

cause
Character (other classes are converted with as.character). The cause for predicting the cause-specific cumulative incidence function in competing risk models, that is the absolute risk of cause events between time zero and times .

...
Only for compatibility reasons.

Details

Predicted (survival) probabilities are returned that can be plotted, summarized and used for inverse of probability of censoring weighting.

Value

type=="surv" A list or a matrix with survival probabilities for all times and all newdata.

type=="cuminc" A list or a matrix with cumulative incidences for all times and all newdata.

type=="list" A list with the following components:

times
The argument times carried forward

predictors
The relevant part of the argument newdata.

indices
A list with the following components

time: Where to find values corresponding to the requested times strata: Where to find values corresponding to the values of the variables in newdata. Together time and strata show where to find the predicted probabilities.

dimensions
a list with the following components: time : The length of times strata : The number of rows in newdata names.strata : Labels for the covariate values.

Author(s)

Thomas Alexander Gerds <tag@biostat.ku.dk>

See Also

predictSurvIndividual

Examples

dat <- SimSurv(400)
fit <- prodlim(Hist(time,status)~1,data=dat)

## predict the survival probs at selected times 
predict(fit,times=c(3,5,10))

## NA is returned when the time point is beyond the
## range of definition of the Kaplan-Meier estimator:
predict(fit,times=c(-1,0,10,100,1000,10000))

## when there are strata, newdata is required 
## or neighborhoods (i.e. overlapping strata)
mfit <- prodlim(Hist(time,status)~X1+X2,data=dat)
predict(mfit,times=c(-1,0,10,100,1000,10000),newdata=dat[18:21,])

## this can be requested in matrix form
predict(mfit,times=c(-1,0,10,100,1000,10000),newdata=dat[18:21,],mode="matrix")

## and even transposed
predict(mfit,times=c(-1,0,10,100,1000,10000),newdata=dat[18:21,],mode="matrix",bytime=TRUE)

[Package prodlim version 2018.04.18 Index]