predictSurvIndividual {prodlim}	R Documentation

Predict individual survival probabilities

Description

Function to extract the predicted probabilities at the individual event times that have been used for fitting a prodlim object.

Usage

predictSurvIndividual(object, lag = 1)

Arguments

object
A fitted object of class "prodlim".

lag
Integer. ‘0’ means predictions at the individual times, 1 means just before the individual times, etc.

Value

A vector of survival probabilities.

Author(s)

Thomas A. Gerds tag@biostat.ku.dk

See Also

predict.prodlim,predictSurv,

Examples

 SurvFrame <- data.frame(time=1:10,status=rbinom(10,1,.5))
 x <- prodlim(formula=Hist(time=time,status!=0)~1,data=SurvFrame)
 predictSurvIndividual(x,lag=1)

[Package prodlim version 2018.04.18 Index]