redist {prodlim}	R Documentation

Calculation of Efron's re-distribution to the right algorithm to obtain the Kaplan-Meier estimate.

Description

Calculation of Efron's re-distribution to the right algorithm to obtain the Kaplan-Meier estimate.

Usage

redist(time, status)

Arguments

time
A numeric vector of event times.

status
The event status vector takes the value 1 for observed events and the value 0 for right censored times.

Value

Calculations needed to

Author(s)

Thomas A. Gerds <tag@biostat.ku.dk>

See Also

prodlim

Examples

redist(time=c(.35,0.4,.51,.51,.7,.73),status=c(0,1,1,0,0,1))

[Package prodlim version 2018.04.18 Index]