summary.Hist {prodlim}	R Documentation

Summary of event histories

Description

Describe events and censoring patterns of an event history.

Usage

## S3 method for class 'Hist'
summary(object, verbose = TRUE, ...)

Arguments

object
An object with class ‘Hist’ derived with Hist

verbose
Logical. If FALSE any printing is supressed.

...
Not used

Value

NULL for survival and competing risk models. For other multi-state models, it is a list with the following entries:

states
the states of the model

transitions
the transitions between the states

trans.frame
a data.frame with the from and to states of the transitions

Author(s)

Thomas A. Gerds tag@biostat.ku.dk

See Also

Hist, plot.Hist

Examples

icensFrame <- data.frame(L=c(1,1,3,4,6),R=c(2,NA,3,6,9),event=c(1,1,1,2,2))
with(icensFrame,summary(Hist(time=list(L,R))))

[Package prodlim version 2018.04.18 Index]