survModel {prodlim}	R Documentation

Survival model for simulation

Description

Create a survival model to simulate a right censored event time data without covariates

Usage

survModel()

Details

This function requires the lava package.

Value

A structural equation model initialized with three variables: the latent event time, the latent right censored time, and the observed right censored event time.

Author(s)

Thomas A. Gerds <tag@biostat.ku.dk>

[Package prodlim version 2018.04.18 Index]