confInt {prodlim}	R Documentation

Add point-wise confidence limits to the graphs of Kaplan-Meier and Aalen-Johansen estimates of survival and cumulative incidence.

Description

This function is invoked and controlled by plot.prodlim.

Usage

confInt(x, times, newdata, type, citype, cause, col, lty, lwd, density = 55,
  ...)

Arguments

x
an object of class ‘prodlim’ as returned by the prodlim function.

times
where to compute point-wise confidence limits

newdata
see plot.prodlim

type
Either "cuminc" or "survival" passed to summary.prodlim as surv=ifelse(type=="cuminc",FALSE,TRUE).

citype
If "shadow" then confidence limits are drawn as colored shadows. Otherwise, dotted lines are used to show the upper and lower confidence limits.

cause
see plot.prodlim

col
the colour of the lines.

lty
the line type of the lines.

lwd
the line thickness of the lines.

density
For citype="shadow", the density of the shade. Default is 55 percent.

...
Further arguments that are passed to the function segments if type=="bars" and to lines else.

Details

This function should not be called directly. The arguments can be specified as Confint.arg in the call to plot.prodlim.

Value

Nil

Author(s)

Thomas Alexander Gerds <tag@biostat.ku.dk>

See Also

plot.prodlim, atRisk, markTime

[Package prodlim version 2018.04.18 Index]